# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flights API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/flights": {
            "post": {
                "tags": [
                    "Flights"
                ],
                "summary": "Get Flights",
                "description": "Fetches a list of flights based on search parameters.",
                "operationId": "getFlights",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "searchs",
                                    "qtyPassengers",
                                    "adult",
                                    "itinerary"
                                ],
                                "properties": {
                                    "searchs": {
                                        "description": "Number of searches",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "qtyPassengers": {
                                        "description": "Quantity of passengers",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "adult": {
                                        "description": "Number of adults",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "itinerary": {
                                        "description": "List of itinerary segments",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "departureCity",
                                                "arrivalCity",
                                                "hour"
                                            ],
                                            "properties": {
                                                "departureCity": {
                                                    "description": "City of departure",
                                                    "type": "string",
                                                    "example": "JFK"
                                                },
                                                "arrivalCity": {
                                                    "description": "City of arrival",
                                                    "type": "string",
                                                    "example": "LAX"
                                                },
                                                "hour": {
                                                    "description": "Date and time of the flight",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-18T10:30:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "dateOfDeparture": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-09-18"
                                            },
                                            "timeOfDeparture": {
                                                "type": "string",
                                                "format": "time",
                                                "example": "10:30:00"
                                            },
                                            "dateOfArrival": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-09-18"
                                            },
                                            "timeOfArrival": {
                                                "type": "string",
                                                "format": "time",
                                                "example": "13:30:00"
                                            },
                                            "marketingCarrier": {
                                                "type": "string",
                                                "example": "AA"
                                            },
                                            "companyName": {
                                                "type": "string",
                                                "example": "American Airlines"
                                            },
                                            "flightOrtrainNumber": {
                                                "type": "string",
                                                "example": "AA123"
                                            },
                                            "locationId": {
                                                "properties": {
                                                    "departureCity": {
                                                        "type": "string",
                                                        "example": "JFK"
                                                    },
                                                    "arrivalCity": {
                                                        "type": "string",
                                                        "example": "LAX"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Flights data not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Flights data not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error fetching flights"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}